# Optimize short string parsing