We use the same SWAR technique used elsewhere in the repo to check 8 bytes for the end quote in a string, before falling back to `bytes.IndexByte`:

```
name           old time/op    new time/op    delta
CodeDecoder-4    5.30ms ± 3%    5.08ms ± 2%  -4.19%  (p=0.000 n=9+10)

name           old speed      new speed      delta
CodeDecoder-4   366MB/s ± 2%   382MB/s ± 2%  +4.36%  (p=0.000 n=9+10)
```

`IndexByte` is pretty quick, but it still has to load the input byte into an XMM register and shuffle it around before starting the routine, and then it also has various length checks to see whether to engage AVX2 or SSE. You also have the function call to consider.

I think it's worth always checking the first chunk of bytes and paying the minor penalty each time you parse a string — it's common to see short strings in object keys, for example.